Fix #24 피드 - 해시태그 (#25)

+++ back/src/main/java/com/rebu/feed/dto/HashtagCountDto.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class HashtagCountDto {
     private String hashtag;
-    private int cnt;
+    private Long cnt;
 }
